+++ packages/client-ledger/RELEASES.md
@@ -147,3 +147,13 @@ Based on:
 - [typescript v1.0.0-alpha.14] packages/client-ledger
 ### Releases
 - [NPM v1.0.0-alpha.14] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.14 - packages/client-ledger
+
+## 2024-08-04 00:17:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.352.1 (2.385.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.15] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.15] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.15 - packages/client-ledger